add [hook]: useRewards
