+++ src/features/mines/utils/ManageBar.tsx
@@ -3,7 +3,7 @@ import { BigNumber } from '@ethersproject/bignumber'
 import { MinusIcon, PlusIcon } from '@heroicons/react/solid'
 import { i18n } from '@lingui/core'
 import { t } from '@lingui/macro'
-import { SOUL_SUMMONER_ADDRESS, Token } from 'sdk'
+import { CurrencyAmount, getPrice, JSBI, SOUL_SUMMONER_ADDRESS, Token, USD } from 'sdk'
 import AssetInput from 'components/AssetInput'
 import { Button } from 'components/Button'
 import { HeadlessUiModal } from 'components/Modal'
@@ -13,24 +13,26 @@ import Web3Connect from 'components/Web3Connect'
 // import { OLD_FARMS } from 'config/farms'
 import { useMineListItemDetailsModal } from 'features/mines/MineListItemDetails'
 import { setMinesModalOpen } from 'features/mines/minesSlice'
-import { classNames, tryParseAmount } from 'functions'
+import { classNames, tryParseAmount, unwrappedToken } from 'functions'
 import { ApprovalState, useApproveCallback } from 'hooks/useApproveCallback'
 import { useActiveWeb3React } from 'services/web3'
 import { useAppDispatch } from 'state/hooks'
 import { useTransactionAdder } from 'state/transactions/hooks'
 import { useCurrencyBalance } from 'state/wallet/hooks'
 import React, { useState } from 'react'
-
+import { useRewards } from 'hooks/useRewards'
 // import { PairType } from '../enum'
 import { useUserInfo } from '../hooks'
 import useMasterChef from '../hooks/useMasterChef'
 import { useV2PairsWithPrice } from 'hooks/useV2Pairs'
 import { useCurrency } from 'hooks/Tokens'
+import { SOUL_ADDRESS } from 'constants/addresses'
+import CurrencyInputPanel from '../components/CurrencyInputPanel'
 
 // @ts-ignore TYPE NEEDS FIXING
 const ManageBar = ({ farm }) => {
   const dispatch = useAppDispatch()
-  const { account, chainId } = useActiveWeb3React()
+  const { account } = useActiveWeb3React() // 250
   const { setContent } = useMineListItemDetailsModal()
   const [toggle, setToggle] = useState(true)
   const [depositValue, setDepositValue] = useState<string>()
@@ -43,7 +45,7 @@ const ManageBar = ({ farm }) => {
   
   farm.pair.token1 ? new Token(
     // @ts-ignore TYPE NEEDS FIXING
-    chainId || 250,
+    250,
     getAddress(farm.lpToken),
     18,
     'SOUL-LP'
@@ -51,7 +53,7 @@ const ManageBar = ({ farm }) => {
     // farm.pair.type === PairType.KASHI ? 'KMP' : 'SLP'
   ) : new Token(
     // @ts-ignore TYPE NEEDS FIXING
-    chainId || 250,
+    250,
     getAddress(farm.lpToken),
     18,
     'SOUL'
@@ -64,14 +66,13 @@ const ManageBar = ({ farm }) => {
   const parsedDepositValue = tryParseAmount(depositValue, liquidityToken)
   const parsedWithdrawValue = tryParseAmount(withdrawValue, liquidityToken)
   // @ts-ignore TYPE NEEDS FIXING
-  const [approvalState, approve] = useApproveCallback(parsedDepositValue, SOUL_SUMMONER_ADDRESS[chainId])
+  const [approvalState, approve] = useApproveCallback(parsedDepositValue, SOUL_SUMMONER_ADDRESS[250])
   
   let token0 = useCurrency(farm.pair.token0?.id)
   let token1 = useCurrency(farm.pair.token1?.id)
-  
-  let [data] = useV2PairsWithPrice([[token0, token1]])
-  let [state, pair, pairPrice] = data
 
+  let [data] = useV2PairsWithPrice([[token0, token1]])
+  // let [state, pair, pairPrice] = data
 
   const depositError = !parsedDepositValue
     ? 'Enter Amount'
@@ -86,6 +87,29 @@ const ManageBar = ({ farm }) => {
     ? 'Insufficient Balance'
     : undefined
   const isWithdrawValid = !withdrawError
+
+  const soulPrice = getPrice(SOUL_ADDRESS[250])
+
+  // const balanceFiatValueRaw
+  //   = pair?.token1 ? Number(pairPrice) * Number(balance?.toSignificant())
+  //   : Number(soulPrice) * Number(balance?.toSignificant())
+
+  // const stakedAmountFiatValueRaw
+  //   = pair?.token1 ? Number(pairPrice) * Number(stakedAmount?.toSignificant())
+  //   : Number(soulPrice) * Number(stakedAmount?.toSignificant())
+
+  // const balanceFiatValue
+  //   = CurrencyAmount.fromRawAmount(
+  //     USD[250],
+  //     JSBI.BigInt(balanceFiatValueRaw.toFixed(USD[250].decimals).toBigNumber(USD[250].decimals))
+  //   )
+
+  // const stakedAmountFiatValue
+  //   = CurrencyAmount.fromRawAmount(
+  //     USD[250],
+  //     JSBI.BigInt(stakedAmountFiatValueRaw.toFixed(USD[250].decimals).toBigNumber(USD[250].decimals))
+  //   )
+// const rewards = useRewards()
 
   return (
     <>
@@ -163,7 +187,7 @@ const ManageBar = ({ farm }) => {
         ( farm.pair.token1 ? (
           <Button
             fullWidth
-            color={!isDepositValid && !!parsedDepositValue ? 'red' : 'blue'}
+            color={!isDepositValid && !!parsedDepositValue ? 'red' : 'purple'}
             onClick={async () => {
               try {
                 // KMP decimals depend on asset, SLP is always 18
@@ -174,7 +198,7 @@ const ManageBar = ({ farm }) => {
                     <HeadlessUiModal.SubmittedModalContent
                       txHash={tx?.hash}
                       header={i18n._(t`Success!`)}
-                      subheader={i18n._(t`Success! Transaction successfully submitted`)}
+                      subheader={i18n._(t`Success! Transaction successfully submitted.`)}
                       onDismiss={() => dispatch(setMinesModalOpen(false))}
                     />
                   )
@@ -193,7 +217,7 @@ const ManageBar = ({ farm }) => {
           ) : (
           <Button
             fullWidth
-            color={!isDepositValid && !!parsedDepositValue ? 'red' : 'blue'}
+            color={!isDepositValid && !!parsedDepositValue ? 'red' : 'purple'}
             onClick={async () => {
               try {
                 // KMP decimals depend on asset, SLP is always 18
@@ -204,7 +228,7 @@ const ManageBar = ({ farm }) => {
                     <HeadlessUiModal.SubmittedModalContent
                       txHash={tx?.hash}
                       header={i18n._(t`Success!`)}
-                      subheader={i18n._(t`Success! Transaction successfully submitted`)}
+                      subheader={i18n._(t`Success! Transaction successfully submitted.`)}
                       onDismiss={() => dispatch(setMinesModalOpen(false))}
                     />
                   )
@@ -239,7 +263,7 @@ const ManageBar = ({ farm }) => {
                   <HeadlessUiModal.SubmittedModalContent
                     txHash={tx?.hash}
                     header={i18n._(t`Success!`)}
-                    subheader={i18n._(t`Success! Transaction successfully submitted`)}
+                    subheader={i18n._(t`Success! Transaction successfully submitted.`)}
                     onDismiss={() => dispatch(setMinesModalOpen(false))}
                   />
                 )
@@ -269,7 +293,7 @@ const ManageBar = ({ farm }) => {
               <HeadlessUiModal.SubmittedModalContent
                 txHash={tx?.hash}
                 header={i18n._(t`Success!`)}
-                subheader={i18n._(t`Success! Transaction successfully submitted`)}
+                subheader={i18n._(t`Success! Transaction successfully submitted.`)}
                 onDismiss={() => dispatch(setMinesModalOpen(false))}
               />
             )
