+++ src/hooks/index.ts
@@ -9,3 +9,4 @@ export { default as useUSDCPrice } from './useUSDCPrice'
 export * from './usePrice'
 export * from './useFarmRewards'
 export * from './useSummonerInfo'
+export * from './useRewards'
